Add Raw versions of all the sockets.
Now there is a socket class hierarchy.  `Socket` is base of the
hierarchy, with two children classes, `RawSocket` and `CookedSocket`.
The normal sockets' names haven't changed, but all the raw versions have
the name `Raw` prefixed.  So for example, Req0 hasn't changed at all,
but now RawReq0 exists as well.

We still need docs and tests.